cmd/go/internal/load: split test logic out of pkg.go into test.go
It's going to grow.

Reviewed-on: https://go-review.googlesource.com/107915
Run-TryBot: Russ Cox <rsc@golang.org>
